Move Bable node creation out of generator.js
