+++ src/serializer/ResidualHeapValueIdentifiers.js
@@ -12,7 +12,7 @@
 import { Value } from "../values/index.js";
 import type { BabelNodeIdentifier } from "@babel/types";
 import invariant from "../invariant.js";
-import type { NameGenerator, PreludeGenerator } from "../utils/generator";
+import type { PreludeGenerator, NameGenerator } from "../utils/PreludeGenerator.js";
 import * as t from "@babel/types";
 
 // This class maintains a map of values to babel identifiers.
