+++ src/utils/PreludeGenerator.js
@@ -1,0 +1,147 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+/* @flow */
+
+import * as t from "@babel/types";
+import { memberExpressionHelper } from "./babelhelpers.js";
+import type {
+  BabelNodeIdentifier,
+  BabelNodeThisExpression,
+  BabelNodeStatement,
+  BabelNodeMemberExpression,
+} from "@babel/types";
+import invariant from "../invariant.js";
+
+export class PreludeGenerator {
+  constructor(debugNames: ?boolean, uniqueSuffix: ?string) {
+    this.prelude = [];
+    this.memoizedRefs = new Map();
+    this.nameGenerator = new NameGenerator(new Set(), !!debugNames, uniqueSuffix || "", "_$");
+    this.usesThis = false;
+    this.declaredGlobals = new Set();
+    this.nextInvariantId = 0;
+  }
+
+  prelude: Array<BabelNodeStatement>;
+  memoizedRefs: Map<string, BabelNodeIdentifier>;
+  nameGenerator: NameGenerator;
+  usesThis: boolean;
+  declaredGlobals: Set<string>;
+  nextInvariantId: number;
+
+  createNameGenerator(prefix: string): NameGenerator {
+    return new NameGenerator(
+      this.nameGenerator.forbiddenNames,
+      this.nameGenerator.debugNames,
+      this.nameGenerator.uniqueSuffix,
+      prefix
+    );
+  }
+
+  convertStringToMember(str: string): BabelNodeIdentifier | BabelNodeThisExpression | BabelNodeMemberExpression {
+    return str
+      .split(".")
+      .map(name => {
+        if (name === "global") {
+          return this.memoizeReference(name);
+        } else if (name === "this") {
+          return t.thisExpression();
+        } else {
+          return t.identifier(name);
+        }
+      })
+      .reduce((obj, prop) => t.memberExpression(obj, prop));
+  }
+
+  globalReference(key: string, globalScope: boolean = false): BabelNodeIdentifier | BabelNodeMemberExpression {
+    if (globalScope && t.isValidIdentifier(key)) return t.identifier(key);
+    return memberExpressionHelper(this.memoizeReference("global"), key);
+  }
+
+  memoizeReference(key: string): BabelNodeIdentifier {
+    let ref = this.memoizedRefs.get(key);
+    if (ref) return ref;
+
+    let init;
+    if (key.includes("(") || key.includes("[")) {
+      // Horrible but effective hack:
+      // Some internal object have intrinsic names such as
+      //    ([][Symbol.iterator]().__proto__.__proto__)
+      // and
+      //    RegExp.prototype[Symbol.match]
+      // which get turned into a babel node here.
+      // TODO: We should properly parse such a string, and memoize all references in it separately.
+      // Instead, we just turn it into a funky identifier, which Babel seems to accept.
+      init = t.identifier(key);
+    } else if (key === "global") {
+      this.usesThis = true;
+      init = t.thisExpression();
+    } else {
+      let i = key.lastIndexOf(".");
+      if (i === -1) {
+        init = t.memberExpression(this.memoizeReference("global"), t.identifier(key));
+      } else {
+        init = t.memberExpression(this.memoizeReference(key.substr(0, i)), t.identifier(key.substr(i + 1)));
+      }
+    }
+    ref = t.identifier(this.nameGenerator.generate(key));
+    this.prelude.push(t.variableDeclaration("var", [t.variableDeclarator(ref, init)]));
+    this.memoizedRefs.set(key, ref);
+    return ref;
+  }
+}
+
+function escapeInvalidIdentifierCharacters(s: string): string {
+  let res = "";
+  for (let c of s)
+    if ((c >= "0" && c <= "9") || (c >= "a" && c <= "z") || (c >= "A" && c <= "Z")) res += c;
+    else res += "_" + c.charCodeAt(0);
+  return res;
+}
+
+const base62characters = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+function base62encode(n: number): string {
+  invariant((n | 0) === n && n >= 0);
+  if (n === 0) return "0";
+  let s = "";
+  while (n > 0) {
+    let f = n % base62characters.length;
+    s = base62characters[f] + s;
+    n = (n - f) / base62characters.length;
+  }
+  return s;
+}
+
+export class NameGenerator {
+  constructor(forbiddenNames: Set<string>, debugNames: boolean, uniqueSuffix: string, prefix: string) {
+    this.prefix = prefix;
+    this.uidCounter = 0;
+    this.debugNames = debugNames;
+    this.forbiddenNames = forbiddenNames;
+    this.uniqueSuffix = uniqueSuffix;
+  }
+  prefix: string;
+  uidCounter: number;
+  debugNames: boolean;
+  forbiddenNames: Set<string>;
+  uniqueSuffix: string;
+  generate(debugSuffix: ?string): string {
+    let id;
+    do {
+      id = this.prefix + base62encode(this.uidCounter++);
+      if (this.uniqueSuffix.length > 0) id += this.uniqueSuffix;
+      if (this.debugNames) {
+        if (debugSuffix) id += "_" + escapeInvalidIdentifierCharacters(debugSuffix);
+        else id += "_";
+      }
+    } while (this.forbiddenNames.has(id));
+    return id;
+  }
+}
